+++ web-components/src/components/robotoff-contribution-message/robotoff-contribution-message.ts
@@ -133,6 +133,7 @@ export class RobotoffContributionMessage extends LanguageCodesMixin(SignalWatche
    */
   private _fetchDataTask = new Task(this, {
     task: async ([productCode]) => {
+      console.log("Fetching data for product code", productCode, this._languageCodes)
       this.showMessages = {
         [RobotoffContributionType.QUESTIONS]: false,
         [RobotoffContributionType.INGREDIENT_SPELLCHECK]: false,
@@ -173,7 +174,7 @@ export class RobotoffContributionMessage extends LanguageCodesMixin(SignalWatche
         ...insightValues,
       }
     },
-    args: () => [this.productCode, this.languageCodes],
+    args: () => [this.productCode, ...this._languageCodes],
   })
 
   /**