+++ web-components/src/components/robotoff-nutrient-extraction/robotoff-nutrient-extraction-form.ts
@@ -15,7 +15,6 @@ import {
   getTaxonomyNameByLang,
   nutrientTaxonomies,
 } from "../../signals/taxonomies"
-import { getLocale } from "../../localization"
 import {
   getPossibleUnits,
   INSIGHTS_ANNOTATION_SIZE,
@@ -38,7 +37,7 @@ import {
   triggerSubmit,
 } from "../../utils"
 import { nutrientsOrderByCountryCode, sortKeysByNutrientsOrder } from "../../signals/openfoodfacts"
-import { countryCode } from "../../signals/app"
+import { countryCode, languageCode } from "../../signals/app"
 
 import "../shared/loading-button"
 import "../shared/autocomplete-input"
@@ -405,6 +404,11 @@ export class RobotoffNutrientExtractionForm extends LitElement {
     console.log("nutrientsOrder", nutrientsOrder)
     const keySet = new Set<string>()
 
+    // Check if nutriments data is available
+    if (!nutrimentsData.nutriments) {
+      return keySet
+    }
+
     // Process nutrients
     INSIGHTS_ANNOTATION_SIZE.forEach((size) => {
       const suffix = NUTRIENT_SUFFIX[size]
@@ -717,7 +721,7 @@ export class RobotoffNutrientExtractionForm extends LitElement {
   ) {
     const inputName = this.getInputValueName(key, column)
     const value = nutrient?.value.toString() ?? ""
-    const label = getTaxonomyNameByIdAndLang(key, getLocale())
+    const label = getTaxonomyNameByIdAndLang(key, languageCode.get())
     const isHidden = this.inputHiddenBySizeAndNutrientKey[column][key]
 
     return html`
@@ -1036,7 +1040,7 @@ export class RobotoffNutrientExtractionForm extends LitElement {
    * @param alreadyAddedNutrients
    */
   renderAddNutrientRow(alreadyAddedNutrients: string[]) {
-    const lang = getLocale()
+    const lang = languageCode.get()
     const filteredNutrientTaxonomies = nutrientTaxonomies
       .get()
       // filter out the nutrients that are already added to the table